+++ lib/form-view/autovalidate/autovalidate.js
@@ -1,7 +1,7 @@
 import o from 'component-dom'
 import debug from 'debug'
 import serialize from 'get-form-data'
-import validate from '../../validate/validate.js'
+import validate from '../validate/validate'
 
 let log = debug('democracyos:autovalidate')
