+++ lib/form-view/form-view.js
@@ -1,7 +1,7 @@
 import o from 'component-dom'
 import spin from 'component-spin'
 import t from 't-component'
-import * as validate from '../validate/validate.js'
+import * as validate from './validate/validate'
 import autosubmit from './autosubmit/autosubmit'
 import autovalidate from './autovalidate/autovalidate'
 import isMobile from '../is-mobile/is-mobile.js'
