+++ lib/form-view/validate/validate.js
@@ -3,7 +3,7 @@
  */
 
 import o from 'component-dom'
-import validators from '../validators/validators.js'
+import validators from '../../validators/validators'
 import debug from 'debug'
 
 let log = debug('democracyos:validate')
